Add repr, formatters and refactor necessary files to work with them

+++ test/event.test.cpp
@@ -8,5 +8,5 @@ using namespace sage;
 TEST_CASE ("Event") {
 	MESSAGE(Event::make_window_closed());
 	for (auto i = 0; i < 10; ++i)
-		MESSAGE(Event::make_window_resized(Size<int>{std::rand() % 4000, std::rand() % 4000}));
+		MESSAGE(Event::make_window_resized(Size<size_t>{std::rand() % 4000, std::rand() % 4000}));
 }
